Refactor couple routes and deleted example and test routes
