--- routes/index.ts
@@ -1,8 +1,0 @@
-import express from "express";
-
-export const indexRouter = express.Router();
-
-/* GET home page. */
-indexRouter.get('/', function(req, res, next) {
-  res.send('index');
-});
