+++ routes/user.ts
@@ -1,5 +1,7 @@
 import express from "express";
-import { getAllEmployers,getUserDataAndSalary, addHours,addPermamentSalary, getUserHistory, paymentRequest,paymentDone,editUser,deleteUser} from "../controllers/userController";
+import { getAllEmployers,getUserDataAndSalary, addHours,addPermamentSalary, getUserHistory, paymentRequest,paymentDone,editUser,deleteUser, setHourSalary, getUnpaid} from "../controllers/userController";
+import {authenticateToken} from "../middleware/authMiddleware";
+
 
 export const userRouter = express.Router();
 
@@ -12,4 +14,7 @@ userRouter.get('/paymentdone/:employerId/:employeeId',paymentDone)
 userRouter.get('/getuserdata/:userid', getUserDataAndSalary);
 userRouter.put('/edituser/:userid', editUser);
 userRouter.delete('/deleteuser/:userId', deleteUser);
+userRouter.post('/setHourSalary', authenticateToken, setHourSalary);
+userRouter.get('/getUnpaid', authenticateToken, getUnpaid)
+
