--- routes/users.ts
@@ -1,8 +1,0 @@
-import express from "express";
-
-export const usersRouter = express.Router();
-
-/* GET users listing. */
-usersRouter.get('/', function(req, res, next) {
-  res.send('respond with a resource');
-});
